* added more tests cases, texture check isn't working (at least for my driver, i suspect a bug in the gl implementation)  * added shader preprocessor from Horazont, not working yet, need to adapt cmake lists  * added second texture, todo more test cases  * some minor fixes due to static code analysis

+++ src/sdlwrapper/sdlwindow.cpp
@@ -68,8 +68,8 @@ void SDLWindow::handleEvent(SDL_Event event) {
     if (event.window.event == SDL_WINDOWEVENT_RESIZED) {
       SDL_Log("resized window");
       if (m_initializedgl) {
-        m_width = event.window.data1;
-        m_height = event.window.data2;
+        m_width = (uint32_t) event.window.data1;
+        m_height = (uint32_t) event.window.data2;
         onResize(m_width, m_height);
       }
     }
